Add Onboarding Fields to Firestore
Also check to see if fields like phone numbers, date of birth, and height inches are the correct format.

The required fields include general section, emergency contact section, and the height from the medical section.

The Cancer, Diabetes, Surgery, Other Conditions, and Allergies fields in the medical section are not required to submit the onboarding form