Update General Information page design + show onboarding required fields
Required onboarding fields appear with an * before the question and with the phrase "(Required)" after the question.

A patient's height is saved to firestore with the units "feet" and "inches" and is shown on the general information page as so.

The general and emergency contact text views are updated to show long texts.