Update Patient Dashboard Activity and Patient General Info Activity
I updated the variables to get general and emergency contact information from firestore on the dashboard. Before, these variables were public static String of length 1 now they are public static string. They are static on the Dashboard activity because I pull the data on the dashboard to show on the general information  general information activity.

Initially, I pulled the data from firestore on the general information activity but it would not update the UI components in time to show it to the user. The code to pull data from the general information page is still included in the code to use if we can find a solution to this issue, those variables are not static.